backend/pyproject.toml
[tool.poetry]
name = "kidofood-backend"
version = "0.1.0"
description = "Backend code for kidofood food deliver web application"
authors = ["noaione <noaione0809@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.86.0"
redis = {extras = ["hiredis"], version = "^4.3.4"}
python-dotenv = "^0.21.0"
orjson = "^3.8.1"
websockets = "^10.4"
coloredlogs = "^15.0.1"
motor = "3.0.0"
pendulum = "^2.1.2"
beanie = "^1.15.3"
argon2-cffi = "^21.3.0"
minio = "^7.1.12"
aiopath = "0.5.12"
itsdangerous = "^2.1.2"
strawberry-graphql = {extras = ["fastapi"], version = "^0.142.2"}

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
isort = "^5.10.1"
black = "^22.10.0"
uvicorn = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py38", "py39"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | env
  | _build
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
extend_skip = ["_dataset/**", "env/**", "venv/**", "./venv/**", ".pytest_cache/**", ".nox/**"]
line_length = 110
skip_gitignore = true

[tool.pyright]
include = ["internals", "pipelines", "routes", "app.py"]
exclude = ["venv", "env", "node_modules", ".venv", ".env", ".nox", ".pytest_cache", ".mypy_cache", ".tox", "build", "dist", "_build", "**/__pycache__"]
ignore = ["_testing"]

reportMissingImports = true
venv = "."

pythonVersion = "3.9"
